Framework for client complete

Approved-by: Diana M Aschner

+++ tenmo-server/src/main/java/com/techelevator/tenmo/controller/TransferController.java
@@ -24,25 +24,25 @@ public class TransferController {
 
     //list transfers by user
     @RequestMapping(path = "/user/{userId}", method = RequestMethod.GET)
-    public List<Transfer> listTransfersByUser(@PathVariable Long userId) {
+    public List<Transfer> listTransfersByUser(@PathVariable Integer userId) {
         return transferDao.listTransfersByUser(userId);
     }
 
     //get transfer by transfer id
     @RequestMapping(path = "/{transferId}", method = RequestMethod.GET)
-    public Transfer getTransferById(@PathVariable Long transferId) {
+    public Transfer getTransferById(@PathVariable Integer transferId) {
         return transferDao.getTransferById(transferId);
     }
 
     //send
     @RequestMapping(path = "/send/{userId}", method = RequestMethod.PUT)
-    public void send(@RequestBody Long transferId){
+    public void send(@RequestBody Integer transferId){
         transferDao.send(transferId);
     }
 
     //withdraw
     @RequestMapping(path = "/withdraw/{userId}", method = RequestMethod.PUT)
-    public void withdraw(@RequestBody Long transferId){
+    public void withdraw(@RequestBody Integer transferId){
         transferDao.withdraw(transferId);
     }
